% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkg_versions.R
\name{add_pkg_versions}
\alias{add_pkg_versions}
\title{Add curent version string to package dependencies}
\usage{
add_pkg_versions(
  pkg = ".",
  fields = c("Depends", "Imports", "LinkingTo", "Suggests")
)
}
\arguments{
\item{pkg}{package description, can be path or package name}

\item{fields}{fields to get & report dependencies for}
}
\description{
Will \code{cat} out a cut/paste-able set of fields for a
\code{DESCRIPTION} file with minimum required versions for
each package based upon currently available package vesions
in CRAN.  https://stackoverflow.com/questions/38738292/how-can-i-automatically-add-update-depends-imports-suggests-versions-in-descript
}
\note{
R and the R version is NOT added to \code{Depends}
}
\examples{
add_pkg_versions("qmethod")
add_pkg_versions("MASS")
\dontrun { # assumes you're in a pkg devel dir
add_pkg_versions()
}
}
