\name{fold.tree}
\alias{fold.tree}

\title{
fold.tree
}
\description{
Implements a recursive tree folding algorthim
}
\usage{
fold.tree(f, x)
}

\arguments{
  \item{f}{
function used to reduce the elements of x
}
  \item{x}{
list of R objects
}
}

\value{
The reduced value of the list x. 
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Eric Kramer
}
\note{
%%  ~~further notes~~
}

