% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_appendix_title}
\alias{fpr_appendix_title}
\title{Build the title for the appendices}
\usage{
fpr_appendix_title(site = my_site, site2 = NULL, site3 = NULL)
}
\arguments{
\item{site}{Numeric value of site.  Usually PSCIS ID. Defaults to object my_site.}

\item{site2}{Numeric value of site 2. Defaults to NULL.}

\item{site3}{Numeric value of site 3. Defaults to NULL.}
}
\description{
Build the title for the appendices
}
