% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_chk_dupes.R
\name{fpr_chk_dupes}
\alias{fpr_chk_dupes}
\title{Check for Duplicate Values in a Column}
\usage{
fpr_chk_dupes(dat = NULL, col_screen = NULL)
}
\arguments{
\item{dat}{[data.frame] A data frame to be checked for duplicates.}

\item{col_screen}{A column in the data frame to check for duplicates. Must be specified as a tidy evaluation argument.}
}
\value{
The function does not return a value but provides a warning if duplicates are found.
}
\description{
This function checks for duplicate values in a specified column of a dataset and provides a warning if duplicates are found.
}
\details{
The function identifies duplicate values in the column and provides a warning with a list of the duplicate values if any are found.
}
