% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fpr_colorize}
\alias{fpr_colorize}
\title{Color font in html resulting from rmarkdown inline text}
\usage{
fpr_colorize(x, color)
}
\arguments{
\item{x}{String of text to colorize.}

\item{color}{String in english}
}
\value{

}
\description{
Color font in html resulting from rmarkdown inline text
}
\examples{
fpr_colorize('color this red', 'red')
}
