% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_dat_w_rows}
\alias{fpr_dat_w_rows}
\title{Test if a dataframe has rows}
\usage{
fpr_dat_w_rows(df)
}
\arguments{
\item{df}{Tibble or dataframe to test to see if it has no rows}
}
\value{
Bolean
}
\description{
Helper function for fpr::fpr_photo_qa()
}
