% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_db_conn.R
\name{fpr_db_conn}
\alias{fpr_db_conn}
\title{Connect to postgres database}
\usage{
fpr_db_conn(
  db_var = Sys.getenv("PG_DB_SHARE"),
  host_var = Sys.getenv("PG_HOST_SHARE"),
  port_var = Sys.getenv("PG_PORT_SHARE"),
  user_var = Sys.getenv("PG_USER_SHARE"),
  password_var = Sys.getenv("PG_PASS_SHARE")
)
}
\arguments{
\item{db_var}{Quoted string value of database name.  Defaults to `Sys.getenv('PG_DB_SHARE')`}

\item{host_var}{Quoted string value of host name.  Defaults to `Sys.getenv('PG_HOST_SHARE')`}

\item{port_var}{Quoted string value of port name.  Defaults to `Sys.getenv('PG_PORT_SHARE')`}

\item{user_var}{Quoted string value of user name.  Defaults to `Sys.getenv('PG_USER_SHARE')`}

\item{password_var}{QQuoted string value of password.  Defaults to `Sys.getenv('PG_PASS_SHARE')`}
}
\description{
Connect to pg database using environmental variables.  Ass
}
\examples{
\dontrun{conn <- fpr_db_conn()}
}
\seealso{
Other database postgres: 
\code{\link{fpr_db_query}()},
\code{\link{fpr_dbq_lstable}()},
\code{\link{fpr_dbq_subset}()}
}
\concept{database postgres}
