% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_db_q_crossings_utm.R
\name{fpr_db_q_crossings_utm}
\alias{fpr_db_q_crossings_utm}
\title{Query crossings by utm zone}
\usage{
fpr_db_q_crossings_utm()
}
\value{
Text string to insert int0 \link{fpr_db_conn} call
}
\description{
Grabs 100 crossings from each of the utm zones for QA
}
\examples{
\dontrun{fpr_db_query(query = fpr_db_q_crossings_utm())}
}
