% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_db_query.R
\name{fpr_db_query}
\alias{fpr_db_query}
\title{Query a postgres database}
\usage{
fpr_db_query(query = "SELECT * FROM bcfishpass.crossings limit 100;", ...)
}
\arguments{
\item{query}{Quoted string query written in `sql`.}
}
\value{
Object pulled from database and message of how long the query took to complete.
}
\description{
Provides connection from \link{fpr_db_conn} as well as a subsequent disconnection
}
\examples{
\dontrun{
fpr_db_query(query = fpr_dbq_subset())
}
}
\seealso{
Other database postgres: 
\code{\link{fpr_db_conn}()},
\code{\link{fpr_dbq_lscols}()},
\code{\link{fpr_dbq_lstables}()},
\code{\link{fpr_dbq_subset}()}
}
\concept{database postgres}
