% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfpr_dbq_clip.R
\name{fpr_dbq_clip}
\alias{fpr_dbq_clip}
\title{Clips a point layer by a polygon layer.}
\usage{
fpr_dbq_clip(schema_table_point, schema_table_polygon, join_column, join_on)
}
\arguments{
\item{schema_table_point}{String (quoted) name of point layer schema.table}

\item{schema_table_polygon}{String (quoted) name of polygon layer schema.table}

\item{join_column}{String (quoted) name of column to join tables on from polygon. See column names of any table with \link{fpr_dbq_lscols}}

\item{join_on}{String (quoted) or vector of specific terms to join on.}
}
\value{
Sf of clipped point layer
}
\description{
Clips a point layer by a polygon layer.
}
\examples{
\dontrun{fpr_db_query(query = fpr_dbq_clip('whse_environmental_monitoring.envcan_hydrometric_stn_sp',
'whse_basemapping.fwa_watershed_groups_poly', 'watershed_group_code', c("FRAN", "NECR")))}
}
\seealso{
Other database postgres: 
\code{\link{fpr_db_conn}()},
\code{\link{fpr_db_query}()},
\code{\link{fpr_dbq_lscols}()},
\code{\link{fpr_dbq_lstables}()},
\code{\link{fpr_dbq_subset}()}
}
\concept{database postgres}
