% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_dbq_lscols.R
\name{fpr_dbq_lscols}
\alias{fpr_dbq_lscols}
\title{List names of all columns in a postgres table}
\usage{
fpr_dbq_lscols(schema = "bcfishpass", table = "crossings_vw", ...)
}
\arguments{
\item{schema}{Character string (quoted) name of target postgres db schema.  See names of schemas with \link{fpr_dbq_lstables}}

\item{table}{Character string (quoted) name of target postgres db table  See names of schemas with \link{fpr_dbq_lstables}}
}
\description{
List names of all columns in a postgres table
}
\examples{
\dontrun{fpr_db_query(query = fpr_dbq_lscols())}
}
\seealso{
Other database postgres: 
\code{\link{fpr_db_conn}()},
\code{\link{fpr_db_query}()},
\code{\link{fpr_dbq_lstables}()},
\code{\link{fpr_dbq_subset}()}
}
\concept{database postgres}
