% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_dbq_lstable.R
\name{fpr_dbq_lstable}
\alias{fpr_dbq_lstable}
\title{List all tables and their sizes}
\usage{
fpr_dbq_lstable(col_order = "table_schema, table_name")
}
\arguments{
\item{col_order}{Single quoted string with columns to order results on. To order by size
use 'pg_relation_size desc'.}
}
\description{
List all tables and their sizes
}
\seealso{
Other database postgres: 
\code{\link{fpr_db_conn}()},
\code{\link{fpr_db_query}()},
\code{\link{fpr_dbq_subset}()}
}
\concept{database postgres}
