% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_dbq_subset.R
\name{fpr_dbq_subset}
\alias{fpr_dbq_subset}
\title{Build query to read in subset table from postgres database.}
\usage{
fpr_dbq_subset(
  col_partition = "utm_zone",
  col_select = "*",
  schema = "bcfishpass",
  table = "crossings",
  num_rows = 100,
  randomize = TRUE,
  col_order = "aggregated_crossings_id",
  ...
)
}
\arguments{
\item{col_partition}{String (quoted) name of column to partition table on. See column names of any table with \link{fpr_dbq_lscols}}

\item{col_select}{String (quoted) name of columns to select from the table.  Must include the `col_order` and `col_partition`.}

\item{schema}{Character string (quoted) name of target postgres db schema.  See names of schemas with \link{fpr_dbq_lstables}}

\item{table}{Character string (quoted) name of target postgres db table  See names of schemas with \link{fpr_dbq_lstables}}

\item{num_rows}{Number of rows to slice (if available) from each partitioned section of table.}

\item{randomize}{Logical (TRUE or FALSE) of whether to choose rows within partitioned subsets optionally.
If FALSE `col_order` needs to be specified}

\item{col_order}{String (quoted) name of column to order partitioned table subsets before sliced.
See column names of any table with \link{fpr_dbq_lscols}}
}
\value{
Text string to insert into \link{fpr_db_conn} call
}
\description{
Build query to read in subset table from postgres database.
}
\examples{
\dontrun{fpr_db_query(query = fpr_dbq_subset())}
}
\seealso{
Other database postgres: 
\code{\link{fpr_db_conn}()},
\code{\link{fpr_db_query}()},
\code{\link{fpr_dbq_clip}()},
\code{\link{fpr_dbq_lscols}()},
\code{\link{fpr_dbq_lstables}()}
}
\concept{database postgres}
