% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fpr_gpkg_p}
\alias{fpr_gpkg_p}
\title{Make point geopackage layer from dataframe}
\usage{
fpr_gpkg_p(
  dat,
  gpkg_name = "mapping",
  utm_zone = 10,
  easting = "utm_easting",
  northing = "utm_northing",
  directory = "data/mapping/"
)
}
\arguments{
\item{dat}{Dataframe with coordinates in}

\item{gpkg_name}{String name of gpkg}

\item{utm_zone}{Numeric value for utm zone}

\item{easting}{Numeric value for}

\item{northing}{Numeric value for}

\item{directory}{String value for}
}
\description{
Make point geopackage layer from dataframe
}
