% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fpr_import_hab_con}
\alias{fpr_import_hab_con}
\title{Import provincial fish data submission template}
\usage{
fpr_import_hab_con(path = "./data/habitat_confirmations.xls")
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Includes habitat data and fish sampling information
}
