% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fpr_import_hab_prior}
\alias{fpr_import_hab_prior}
\title{Import habitat priorities csv and assign aggregated crossing id}
\usage{
fpr_import_hab_prior(
  input = "data/habitat_confirmations_priorities.csv",
  filter_ef_s2 = TRUE
)
}
\arguments{
\item{input}{csv to input which has site as a column}

\item{filter_ef_s2}{Boolean - Do we filter out electrofishing sites, feature_record_only ie.falls, LWD jams, and sites with multiple cards for different sections surveyed?}
}
\value{
dataframe of the priorities with an aggregated_crossings_id added
}
\description{
Import habitat priorities csv and assign aggregated crossing id
}
