% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fpr_import_pscis}
\alias{fpr_import_pscis}
\title{Import the pscis template}
\usage{
fpr_import_pscis(workbook_name = "pscis_phase1.xlsm", dir_root = "data", ...)
}
\arguments{
\item{workbook_name}{string value for name of worksheet}

\item{dir_root}{String indicating name of directory to look for pscis submission file}

\item{...}{Unused - Pass through another param to \link{fpr_sheet_trim}}
}
\value{
A dataframe
}
\description{
Import the pscis template. Assigns the column types based on the fpr_xref_pscis data object.
Extracts a time_start and a date_time_start of the survey from the comments if it is sritten in 24hr format after the last period of the comments column.
Assigns the camera tot hte first person listed in the crew memebers column. Creates a site_id as either the PSCIS id or the my crossing reference
 This is a helper for fpr_import_pscis_all.
}
