% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_kable}
\alias{fpr_kable}
\title{Custom kable settings}
\usage{
fpr_kable(
  dat,
  caption_text = "",
  font = font_set,
  footnote_text = NULL,
  scroll = TRUE
)
}
\arguments{
\item{dat}{dataframe to make a table with}

\item{caption_text}{string to insert as caption}

\item{font}{size of font.  Defaults to font_set usually set in setup chunks of index.Rmd based on output type ie. gitbook vs pagedown}

\item{footnote_text}{string to insert as footnote}

\item{scroll}{TRUE or FALSE about whether to have scroll}
}
\value{

}
\description{
Custom kable settings
}
\examples{
fpr_kable(mtcars)
}
