% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{fpr_make_html_tbl}
\alias{fpr_make_html_tbl}
\title{Make html table from dataframe}
\usage{
fpr_make_html_tbl(dat)
}
\arguments{
\item{dat}{Dataframe or tibble which usually only has one row}
}
\value{
html object
}
\description{
Usually just a one row tibble is transformed as this is helper function for
making kml files from PSCIS spreadsheets.  Other functions for the kml are in the 0110-make-kml.R file
in the fish_passage_skeena_2021_reporting repo for now.
}
\seealso{
Other spatial: 
\code{\link{fpr_sp_wshd_stats}()}
}
\concept{spatial}
