% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_cost_estimate}
\alias{fpr_my_cost_estimate}
\title{Sentence to explain cost to fix site}
\usage{
fpr_my_cost_estimate(
  dat = tab_cost_est_phase2,
  col_filter = pscis_crossing_id,
  site = my_site,
  col_pull = cost_est
)
}
\arguments{
\item{dat}{Dataframe with pscis_crossing_id and cost_est_1000s columns}

\item{col_filter}{String (unqouted) name of column to filter. Defaults to pscis_crossing_id}

\item{site}{Integer. Corresponds to the PSCIS site id. Defaults to my_site defined in environment}
}
\value{
String. Formatted number
}
\description{
Sentence to explain cost to fix site
}
