% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_cover_sentence}
\alias{fpr_my_cover_sentence}
\title{Sentence to describe the dominant and subdominant cover types}
\usage{
fpr_my_cover_sentence(loc = "us", sit = my_site)
}
\arguments{
\item{loc}{String in quotes which is either 'us' upstream or 'ds' downstream.  Defaults to 'us'}

\item{sit}{Integer. Corresponds to the PSCIS site id. Defaults to my_site defined in environment}
}
\value{
String sentence with the cover types described
}
\description{
Sentence to describe the dominant and subdominant cover types
}
\examples{
fpr_my_cover_sentence()
}
