% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_habitat_cover}
\alias{fpr_my_habitat_cover}
\title{Pull out dominant and subdominant cover types}
\usage{
fpr_my_habitat_cover(
  dat = hab_site,
  sit = my_site,
  loc = "us",
  cover_type = "dominant"
)
}
\arguments{
\item{dat}{Dataframe. Defaults to hab_site.}

\item{sit}{Integer. Corresponds to the PSCIS site id. Defaults to my_site defined in envrionment}

\item{loc}{String. Either 'us' (upstream) or 'ds' (downstream). Use quotes.}

\item{cover_type}{Cover type from the fish submission spreadsheet}
}
\value{
string value of cover as per submission template dropdown options
}
\description{
Pull out dominant and subdominant cover types
}
