% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_habitat_info}
\alias{fpr_my_habitat_info}
\title{Extract habitat details from hab_site dataframe}
\usage{
fpr_my_habitat_info(
  dat = hab_site,
  sit = my_site,
  loc = "us",
  col_pull = "avg_channel_width_m"
)
}
\arguments{
\item{dat}{Dataframe. Defaults to hab_site.}

\item{sit}{Integer. Corresponds to the PSCIS site id. Defaults to my_site defined in envrionment}

\item{loc}{String. Either 'us' (upstream) or 'ds' (downstream). Use quotes.}

\item{row}{String in quotes indicating name of column (actually a row at this point) to pull out.}
}
\value{
Sring or number from hab data
}
\description{
Extract habitat details from hab_site dataframe
}
