% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_mapsheet}
\alias{fpr_my_mapsheet}
\title{Print link to map based on watershed name and 50k grid found in bcfishpass}
\usage{
fpr_my_mapsheet(wshd = "bulkley")
}
\arguments{
\item{wshd}{String identifying the watershed.  Not standardized. Must match folders at https://hillcrestgeo.ca/outgoing/fishpassage/projects/}
}
\value{
Stirng with the url for the map where the crossing is located
}
\description{
Print link to map based on watershed name and 50k grid found in bcfishpass
}
