% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_priority_info}
\alias{fpr_my_priority_info}
\title{Extract info from habitat_confirmations_priorities spreadsheet}
\usage{
fpr_my_priority_info(
  dat = habitat_confirmations_priorities,
  sit = my_site,
  loc = "us",
  col_pull = hab_value
)
}
\arguments{
\item{dat}{Dataframe. Defaults to habitat_confirmations_priorities}

\item{sit}{Integer. Corresponds to the PSCIS site id. Defaults to my_site defined in envrionment}

\item{loc}{String. Either 'us' (upstream) or 'ds' (downstream). Used quotes.}

\item{col_pull}{String. Tidyselect format. Column name for information you want to pull.}
}
\value{
String
}
\description{
Extract info from habitat_confirmations_priorities spreadsheet
}
\examples{
fsr_my_priority_info()
}
