% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_pscis_info}
\alias{fpr_my_pscis_info}
\title{Pull info about a site}
\usage{
fpr_my_pscis_info(
  dat = pscis_phase2,
  col_filter = pscis_crossing_id,
  site = my_site,
  col_pull = stream_name
)
}
\arguments{
\item{dat}{Dataframe to pull info from. Defaults to object pscis_phase2 built in tables.R}

\item{col_filter}{String name of column on which to filter dat. Defaults to `pscis_crossing_id`.}

\item{site}{Numeric value of site of which to filter dat. Defaults to object `my_site`.}

\item{col_pull}{String name of column of which to pull from filtered `dat`. Defaults to `stream_name`}
}
\description{
Pull info about a site
}
