% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_survey_dist}
\alias{fpr_my_survey_dist}
\title{Sentence to describe survey length of your site.}
\usage{
fpr_my_survey_dist(loc = "us", sit = my_site)
}
\arguments{
\item{loc}{String in quotes which is either 'us' upstream or 'ds' downstream.  Defaults to 'us'}

\item{sit}{Integer. Corresponds to the PSCIS site id. Defaults to my_site defined in environment}
}
\value{
String sentence with the direction and distance surveyed.
}
\description{
No hard stop at the end of the sentence so we can reference the photos.
}
\examples{
fpr_my_survey_dist()
}
