% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_amalg_cv}
\alias{fpr_photo_amalg_cv}
\title{Amalgamate 6 photos into 1}
\usage{
fpr_photo_amalg_cv(site_id)
}
\arguments{
\item{site_id}{Numeric value of site corresponding to the directory name that holds the photos
which include 'road', 'inlet', 'upstream', 'downstream', 'outlet', barrel' in the filenames.}
}
\value{

}
\description{
Here we stack up and down then side to side for reporting.  There must be all 6 photos
present, only 1 of each and within the file names of those photos must be the strings
containing upstream, downstream, inlet, outlet, barrel, road
}
