% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_change_ext}
\alias{fpr_photo_change_ext}
\title{Rename file extensions to what they will be after the change}
\usage{
fpr_photo_change_ext(filenames_to_change)
}
\arguments{
\item{filenames_to_change}{String full path names of photos we want to change}
}
\value{
String of changed file name
}
\description{
Rename file extensions to what they will be after the change
}
