% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_document_all}
\alias{fpr_photo_document_all}
\title{Document all the photos that are in the directories to csv as a bay to track changes after scripted organization}
\usage{
fpr_photo_document_all(path_to_photo_dir, full_names = T)
}
\arguments{
\item{path_to_photo_dir}{Full path to photo directories to be}

\item{full_names}{whether we should use full names when listing photos}
}
\description{
Document all the photos that are in the directories to csv as a bay to track changes after scripted organization
}
