% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_flip}
\alias{fpr_photo_flip}
\title{Flip an image}
\usage{
fpr_photo_flip(site_id = my_site, rotate = 180, str_to_pull = "barrel")
}
\arguments{
\item{site_id}{Numeric indicating site id cooresponding to name of directory containing the images}

\item{rotate}{Numeric value of the amount of rotation in degrees. Defaults to 180}

\item{str_to_pull}{Sting value in photo name.  Must be unique in file}
}
\value{

}
\description{
Flip an image
}
