% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_interval_idx}
\alias{fpr_photo_interval_idx}
\title{Get the index of a time according to which interval it falls within}
\usage{
fpr_photo_interval_idx(date_input, intervs)
}
\arguments{
\item{date_input}{POSIXct date time}

\item{intervs}{Formal class Interval}
}
\value{
Number representing index
}
\description{
For sorting photos into folders based on time they were taken cross referenced to field notes on site cards
}
