% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_paths_to_copy_phase2}
\alias{fpr_photo_paths_to_copy_phase2}
\title{Untested!! Identify photos that should be copied over into a file to upload to the phase 2 PSCIS interface
This is every photo not moved by fpr_photo_paths_to_copy
adapted from https://stackoverflow.com/questions/56160445/how-to-pass-multiple-necessary-patterns-to-str-subset}
\usage{
fpr_photo_paths_to_copy_phase2(path_to_photo_dir)
}
\arguments{
\item{path_to_photo_dir}{Full path to photo directories that will be copied over}
}
\description{
Untested!! Identify photos that should be copied over into a file to upload to the phase 2 PSCIS interface
This is every photo not moved by fpr_photo_paths_to_copy
adapted from https://stackoverflow.com/questions/56160445/how-to-pass-multiple-necessary-patterns-to-str-subset
}
