% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_pull_by_str}
\alias{fpr_photo_pull_by_str}
\title{Pull names of a photo based on a string in the file name}
\usage{
fpr_photo_pull_by_str(site = my_site, str_to_pull = "barrel")
}
\arguments{
\item{site}{Numeric indicating site id cooresponding to name of directory containing the images}

\item{str_to_pull}{String value to search for in name of photo. must be unique in the file. Defaults to barrel}
}
\value{
String representing the full path filename of photo that matches
}
\description{
Pull names of a photo based on a string in the file name
}
