% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_qa}
\alias{fpr_photo_qa}
\title{QA Photos}
\usage{
fpr_photo_qa(dat = pscis_all, col = site_id)
}
\arguments{
\item{dat}{Dataframe to pull site IDs from that coincide with photo folders. Defaults to pscis_phase1}

\item{col}{Column to pull to get site IDs. Defaults to my_crossing_reference}
}
\value{
List of tibbles for each sites.
}
\description{
Ensure that there are 6 named photos and no duplicates named.
}
\details{
find sites with 0 Rows.
fpr::fpr_photo_qa()[fpr::fpr_photo_qa() %>%
map(fpr::fpr_dat_w_rows) %>%
grep(pattern = F) ] %>%
names(.) %>%
unique(.)

See and fix duplicates with
fpr::fpr_photo_qa() %>%
data.table::rbindlist(fill = T)

After dups are fixed then Query for missing values with
dplyr::bind_rows() %>%
dplyr::filter(dplyr::if_any(dplr::everything(), is.na))
}
\examples{
fpr_photo_qa()
}
