% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_qa_df}
\alias{fpr_photo_qa_df}
\title{QA photos to see if all required PSCIS photos are named and look for duplicates of those names.}
\usage{
fpr_photo_qa_df()
}
\description{
Produce data frame showing missing photos and sites with duplicates. Old way of doing it documented here.
Uses helper functions \link{fpr_photo_qa_prep} and \link{fpr_photo_qa}
}
\details{
#' find sites with 0 Rows.
fpr::fpr_photo_qa()[fpr::fpr_photo_qa() %>%
map(fpr::fpr_dat_w_rows) %>%
grep(pattern = F) ] %>%
names(.) %>%
unique(.)


See and fix duplicates with
fpr::fpr_photo_qa() %>%
data.table::rbindlist(fill = T)

After dups are fixed then Query for missing values with
dplyr::bind_rows() %>%
dplyr::filter(dplyr::if_any(dplr::everything(), is.na))
}
