% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_qa_missing_all}
\alias{fpr_photo_qa_missing_all}
\title{Find sites missing all photos}
\usage{
fpr_photo_qa_missing_all(...)
}
\arguments{
\item{...}{Not used. Used to pass `dir_photos = 'filepath'` to \link{fpr_photo_qa}}
}
\value{
Tibble with 7 columns.
}
\description{
Find sites that exist in the PSCIS input spreadsheets in the `data` directory but do not have a directory or do not
have any of the required PSCIS tagged photo names. Fed into \link{fpr_photo_qa_df} and relies on \link{fpr_photo_qa}.
Site directories must be named as numbers.
}
\seealso{
Other photo: 
\code{\link{fpr_photo_amalg_cv}()},
\code{\link{fpr_photo_folders}()}
}
\concept{photo}
