% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_qa_prep}
\alias{fpr_photo_qa_prep}
\title{Helper function to QA Photos}
\usage{
fpr_photo_qa_prep(site_id)
}
\arguments{
\item{site_id}{Numeric value of site corresponding to folder name}
}
\description{
Prep for fpr_photo_qa which will QA photos. Ensure that there are 6 named photos and no duplicates named.
Duplicates do not just contain the keyword but also a preceding underscore and proceeding period (ie. `_keyword.`).
Used inside of
}
