% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_remove_dupes}
\alias{fpr_photo_remove_dupes}
\title{Remove duplicated photos before or after photo name appended to end of photo file name by [fpr_photo_rename()]}
\usage{
fpr_photo_remove_dupes(
  dir_target =
    "/Users/airvine/Library/CloudStorage/OneDrive-Personal/Projects/repo/fish_passage_peace_2023_reporting/data/photos/test/",
  col_time = date_time_original,
  col_photo_name = source_file,
  col_model = model,
  col_iso = iso,
  remove_renamed = FALSE,
  ...
)
}
\arguments{
\item{dir_target}{String. Full path name of directory to be scanned for duplicate photos with different names.}

\item{col_time}{The bare (unquoted) name of the column containing the exif parameter stamping when photo was created.
Defaults to `date_time_original`.}

\item{col_photo_name}{The bare (unquoted) name of the column containing the exif parameter with full path photo name.
Defaults to `source_file`.}

\item{col_model}{The bare (unquoted) name of the column containing the exif parameter stamping camera type that took the photo.
Defaults to `model`.}

\item{col_iso}{he bare (unquoted) name of the column containing the exif parameter stamping photo iso.
Defaults to `iso`.}

\item{remove_renamed}{Logical. Should renamed photos be removed (based on length of photo name)? Defaults to FALSE so that
duplicate photos that have not been renamed are removed.}

\item{...}{Not used.}
}
\description{
Remove duplicated photos before or after photo name appended to end of photo file name by [fpr_photo_rename()]
}
