% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_rename_ext}
\alias{fpr_photo_rename_ext}
\title{Rename photos}
\usage{
fpr_photo_rename_ext(filescopy, filespaste)
}
\arguments{
\item{filescopy}{String full path names of photos we want to change}

\item{filespaste}{String of changed file name}
}
\value{
Action renameing files
}
\description{
Rename photos
}
