% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_resize_batch}
\alias{fpr_photo_resize_batch}
\title{Resize batch of images and convert JPG}
\usage{
fpr_photo_resize_batch(dir_source = NULL, dir_target = NULL, ...)
}
\arguments{
\item{dir_source}{String full path name of directory where photos are at}

\item{dir_target}{String full path name of directory where photos are to be copied to}

\item{...}{Not used. For passing alternative destination photo size as string to \link{fpr_photo_resize_convert}}
}
\value{
Photos resized with `.JPG` extension
}
\description{
Resize batch of images and convert JPG
}
