% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_resize_convert}
\alias{fpr_photo_resize_convert}
\title{Resize images and convert JPG}
\usage{
fpr_photo_resize_convert(photo, size = "1440x1080!", path = NULL)
}
\arguments{
\item{photo}{String of full path name of photo}

\item{size}{String image width and height of photo you want to resize to}

\item{path}{String of path to the directory where you want to burn the converted img}
}
\value{

}
\description{
Resize images and convert JPG
}
