% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_sort_metadat}
\alias{fpr_photo_sort_metadat}
\title{Get photo sorting specific metadata from the photos in the file}
\usage{
fpr_photo_sort_metadat(input_file)
}
\arguments{
\item{input_file}{String full path to directory containing photos we need sorted by time}
}
\value{
A tibble
}
\description{
Get photo sorting specific metadata from the photos in the file
}
