% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_sort_plan}
\alias{fpr_photo_sort_plan}
\title{Make plan for which folder to put each photo in based on surveyor and time}
\usage{
fpr_photo_sort_plan(surveyor)
}
\arguments{
\item{surveyor}{String with the initials of the surveyor input first into spreadsheet because they had the camera. This must
also be the name of the folder in which the raw photos are stored before transfer.}
}
\description{
Time is input into PSCIS submission template in comments after last period
Time is input into habitat_confirmation_priorities.csv template within a time_start column
This function assumes you deine the folders you want to
}
