% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photos_ext_to_change}
\alias{fpr_photos_ext_to_change}
\title{Find photos that should be converted to another file extension}
\usage{
fpr_photos_ext_to_change(target)
}
\arguments{
\item{target}{String that is path to the directory where the photos are that may need renameing.}
}
\value{
String identifying full path of photos to be renamed.
}
\description{
Find photos that should be converted to another file extension
}
