% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fpr_sheet_trim}
\alias{fpr_sheet_trim}
\title{Trim up an excel worksheet - usually during import}
\usage{
fpr_sheet_trim(dat, clean_col2 = FALSE)
}
\arguments{
\item{dat}{Dataframe}

\item{clean_col2}{Logical whether to remove rows if no `gazetted_name` is specified. Defaults to false.}

\item{...}{Not used}
}
\value{
A dataframe
}
\description{
Select column names based on first row with inputs all across.
Trim up the dataframe based on whether there are events in the first column.
This is a `date` in the pscis inputs and `reference_number` in the fish submission template
}
