% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_sp_assign_latlong.R
\name{fpr_sp_assign_latlong}
\alias{fpr_sp_assign_latlong}
\title{Extract latitude and longitude from sf object}
\usage{
fpr_sp_assign_latlong(dat = NULL, col_lat = "lat", col_lon = "lon")
}
\arguments{
\item{dat}{sf dataframe of type `Geometry type: POINT`}

\item{col_lat}{Quoted character string defining the name of the column where the northing will be stored.}

\item{col_lon}{Quoted character string defining the name of the column where the easting will be stored.}
}
\value{
sf object in equivalent CRS as input with lat and lon columns appended.
}
\description{
Extract latitude and longitude (WSG84 CRS 4326 from sf object and add as columns.)
}
\seealso{
Other spatial operations: 
\code{\link{fpr_sp_assign_utm}()}
}
\concept{spatial operations}
