% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_sp_assign_sf_from_utm.R
\name{fpr_sp_assign_sf_from_utm}
\alias{fpr_sp_assign_sf_from_utm}
\title{Assign a spatial data frame from UTM coordinates}
\usage{
fpr_sp_assign_sf_from_utm(
  dat = NULL,
  col_utm_zone = "utm_zone",
  col_easting = "easting",
  col_northing = "northing",
  crs_return = 3005
)
}
\arguments{
\item{dat}{A data frame.}

\item{col_utm_zone}{The name of the column to group by. Default is "utm_zone".}

\item{col_easting}{The name of the easting coordinate column. Default is "easting".}

\item{col_northing}{The name of the northing coordinate column. Default is "northing".}

\item{crs_return}{Numeric. The EPSG code to transform the coordinate reference system to. Must be Geographic (ex. Albers or WSG84). Default is BC Albers - 3005.}
}
\value{
A spatial data frame.
}
\description{
This function is designed to work on a dataframe that contains columns with details of UTM zone,
easting and northing from a NAD 83 UTM Zone coordinate systems. It prepares a spatial data frame
by grouping based on a specified column (UTM zone), converting to a spatial data frame, transforming
the coordinate reference system, and then binding the rows back together. The returned dataframe
must be in a Geographic Coordinate System (BC Albers or vs a Spatial coordinate system like UTM.
}
\examples{
\dontrun{
fpr_sp_assign_sf_from_utm(pscis_all)
}
}
\seealso{
Other spatial operations: 
\code{\link{fpr_sp_assign_latlong}()},
\code{\link{fpr_sp_assign_utm}()},
\code{\link{fpr_sp_gpkg_backup}()}
}
\concept{spatial operations}
