% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_sp_assign_utm.R
\name{fpr_sp_assign_utm}
\alias{fpr_sp_assign_utm}
\title{Import sf object (point) and assign utm zone, easting and northing}
\usage{
fpr_sp_assign_utm(
  dat = NULL,
  col_zone = "utm_zone",
  col_easting = "easting",
  col_northing = "northing",
  sig_dit = 0
)
}
\arguments{
\item{dat}{sf dataframe of type `Geometry type: POINT`}

\item{col_zone}{Quoted string defining the name of the column where the utm zone will be stored.}

\item{col_easting}{Quoted string defining the name of the column where the utm easting will be stored.}

\item{col_northing}{Quoted string defining the name of the column where the utm northing will be stored.}

\item{sig_dit}{Number of significant digits to round the easting and northing to.}
}
\description{
Import sf object (point) and assign utm zone, easting and northing
}
\seealso{
Other spatial operations: 
\code{\link{fpr_sp_assign_latlong}()}
}
\concept{spatial operations}
