% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_sp_utm_assign.R
\name{fpr_sp_utm_assign}
\alias{fpr_sp_utm_assign}
\title{Import sf object (point) and assign utm zone, easting and northing}
\usage{
fpr_sp_utm_assign(
  form = NULL,
  id = site_id,
  col_zone = "utm_zone",
  col_easting = "utm_easting",
  col_northing = "utm_northing"
)
}
\arguments{
\item{form}{sf dataframe of type `Geometry type: POINT`}

\item{id}{Unquoted string (Tidyselect format) that identifies a column with a **unique ID** that can be used to
split the dataframe so coordinates are computed for each row.}

\item{col_zone}{Quoted string defining the name of the column where the utm zone will be stored.}

\item{col_easting}{Quoted string defining the name of the column where the utm easting will be stored.}

\item{col_northing}{Quoted string defining the name of the column where the utm northing will be stored.}
}
\description{
Import sf object (point) and assign utm zone, easting and northing
}
