% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{fpr_sp_watershed}
\alias{fpr_sp_watershed}
\title{Retrieve watershed geometry using fwapg}
\usage{
fpr_sp_watershed(dat)
}
\arguments{
\item{dat}{Dataframe with a stream_crossing_id, blue_line_key and downstream_route_measure defined. Ususally pulled from bcfishpass}
}
\value{
sf object representing watershed areas
}
\description{
Retrieve watershed geometry using fwapg
}
