% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_bcfp}
\alias{fpr_table_bcfp}
\title{Summary table of bcfishpass modelling outputs}
\usage{
fpr_table_bcfp(
  dat = bcfishpass,
  xref_table = xref_bcfishpass_names,
  site = my_site,
  col = stream_crossing_id,
  ...
)
}
\arguments{
\item{dat}{Dataframe with one row. Defaults to bcfishpass}

\item{xref_table}{Table of column names from bcfishpass with reader friendly versions and columns
indicating order and location (left or right) for how to present the information.}

\item{site}{Integer identifying which site to filter the bcfishpass dataframe}

\item{col}{Column to filter bcfishpass.  Defaults to stream_crossing_id}

\item{...}{Not use. Pass param to \link{fpr_kable}}
}
\value{

}
\description{
Built from bcfishpass with the names of the left hand side of the table and 'potential' habitat calculated
}
