% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_bcfp_html}
\alias{fpr_table_bcfp_html}
\title{Write bcfishpass model summaries to html files}
\usage{
fpr_table_bcfp_html(sites, ...)
}
\arguments{
\item{sites}{Integer value representing site.}

\item{...}{Not used. Pass param to fpr_table_bcfp (or through to fpr_kable)}
}
\description{
Tables will be linked to map points
}
