% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_cv_detailed}
\alias{fpr_table_cv_detailed}
\title{Make two column dataframe of PSCIS info with parameter on left and value on right within each column.}
\usage{
fpr_table_cv_detailed(dat)
}
\arguments{
\item{dat}{Datframe of PSCIS spreadsheet input data with only 1 row. Split from pscis_phase1, pscis_reasessments or pscis_phase2.}
}
\value{
Two column dataframe with parameter on left and value on right within each column.
}
\description{
Make two column dataframe of PSCIS info with parameter on left and value on right within each column.
}
