% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_cv_detailed_comments}
\alias{fpr_table_cv_detailed_comments}
\title{Make table for the comments to be appended to}
\usage{
fpr_table_cv_detailed_comments(dat)
}
\arguments{
\item{dat}{Dataframe of PSCIS spreadsheet input data with only 1 row. Split from pscis_phase1, pscis_reasessments or pscis_phase2.}
}
\value{
Dataframe two columns wide with comments: on the left
}
\description{
Make table for the comments to be appended to
}
