% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_cv_detailed_print}
\alias{fpr_table_cv_detailed_print}
\title{Print out a list of fpr_table_cv_detailed tables showing PSCIS data in untidy table in either gitbook or pdf versions.}
\usage{
fpr_table_cv_detailed_print(tab_sum, comments, photos, gitbook_switch = TRUE)
}
\arguments{
\item{tab_sum}{list of dataframes formatted by fpr_table_cv_detailed.}

\item{comments}{list of dataframes formatted by fpr_table_cv_detailed_comments}

\item{photos}{list of dataframes with just the url of the photo for each site}

\item{gitbook}{logical based on whether out put is gitbook or pagedown pdf}
}
\value{
html tables with photo as footnotes
}
\description{
Print out a list of fpr_table_cv_detailed tables showing PSCIS data in untidy table in either gitbook or pdf versions.
}
