% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_cv_html}
\alias{fpr_table_cv_html}
\title{Make html tables that link to interactive map}
\usage{
fpr_table_cv_html(site)
}
\arguments{
\item{site}{Integer. Site ID for site to be summarized.}
}
\value{
html file representing summary table for culvert details
}
\description{
Depends on fpr_table_cv_detailed function and external data xref_names_pscis kept in package.
}
