% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_cv_summary}
\alias{fpr_table_cv_summary}
\title{Make table with summary of culvert details for results section of report}
\usage{
fpr_table_cv_summary(dat)
}
\arguments{
\item{dat}{Dataframe built from fpr_import_pscis}
}
\description{
Depends on tab_culvert_prep object and xref_names objects build in the tables.R file
}
