% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_cv_summary_memo}
\alias{fpr_table_cv_summary_memo}
\title{Summary table of culvert info with photo to place in memos.}
\usage{
fpr_table_cv_summary_memo(
  dat = pscis_phase2,
  site = my_site,
  site_photo_id = my_site,
  font = 11,
  col_filter = pscis_crossing_id
)
}
\arguments{
\item{dat}{Dataframe. Defaults to pscis_phase2}

\item{site}{Integer. Name of PSCIS site to filter pscis_phase2. Defaults to my_site}

\item{site_photo_id}{Integer. Name of PSCIS site to source photo. Defaults to my_site}

\item{font}{Integer. Size of font.}

\item{col_filter}{String unquoted.  Name of column to filter site from. Defaults to pscis_crossing_id}
}
\value{
html object
}
\description{
Summary table of culvert info with photo to place in memos.
}
