% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_fish_density}
\alias{fpr_table_fish_density}
\title{Summary table for fish densities within individual sites}
\usage{
fpr_table_fish_density(dat = fish_abund, sit = my_site, ...)
}
\arguments{
\item{dat}{Dataframe.  Defaults to fish_abund. Must have columns site, local_name, species_code,
life_stage, catch, density_100m2, nfc_pass}

\item{sit}{Integer.  Site ID.  Defaults to my_site}

\item{...}{Not used. Open for passing arguments to \link{fpr_kable}}
}
\value{

}
\description{
Summary table for fish densities within individual sites
}
\seealso{
\link{fpr_kable}
}
