% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_fish_site}
\alias{fpr_table_fish_site}
\title{Summary table for fish sampling sites}
\usage{
fpr_table_fish_site(dat = tab_fish_sites_sum, sit = my_site, ...)
}
\arguments{
\item{dat}{Dataframe.  Defaults to tab_fish_sites_sum.  Must have columns named site, passes, ef_length_m, ef_width_m, area_m2, enclosure}

\item{sit}{Integer.  Site ID.  Defaults to my_site}

\item{...}{Not used. Open for passing arguments to \link{fpr_kable}}
}
\value{
Dataframe
}
\description{
Summary table for fish sampling sites
}
\seealso{
\link{fpr_kable}
}
