% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_table_wshd_sum}
\alias{fpr_table_wshd_sum}
\title{Summary stats for each watershed}
\usage{
fpr_table_wshd_sum(dat = wshds, site_id = NULL)
}
\arguments{
\item{dat}{Dataframe with stream_crossing_id, area_km}

\item{site_id}{Numeric id corresponding to stream_crossing_id (psscis id) of site}
}
\description{
Summary stats for each watershed
}
