% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_test}
\alias{fpr_test}
\title{Print out test table}
\usage{
fpr_test(tab_sum, comments, photos, gitbook_switch = gitbook_on)
}
\arguments{
\item{tab_sum}{Dataframe formatted by fpr_table_cv_detailed.}

\item{comments}{Dataframe formatted by fpr_table_cv_detailed_comments}

\item{photos}{test}

\item{gitbook_switch}{test}
}
\value{
html table with photo as footnotes
}
\description{
Print out test table
}
