% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_xfm_paw_all_scores_result.R
\name{fpr_xfm_paw_all_scores_result}
\alias{fpr_xfm_paw_all_scores_result}
\title{Transform PSCIS Assessment Worksheet Data by Calculating All Metrics, Scores, and Results Necessary for Barrier Risk Assessment}
\usage{
fpr_xfm_paw_all_scores_result(dat)
}
\arguments{
\item{dat}{[dataframe] A dataframe containing the PSCIS data.}
}
\value{
[dataframe] A dataframe with all calculated metrics, scores, and results necessary for barrier risk assessment.
}
\description{
This function calculates the fall metrics, scores, and results for BC Provincial Stream Crossing Inventory System (PSCIS) data,
sums score columns, and assigns `barrier_result`.
}
\examples{
path <- system.file("extdata", "pscis_phase1.xlsm", package = "fpr")
dat <- fpr_import_pscis(dir_root = fs::path_dir(path))
result <- fpr_xfm_paw_all_scores_result(dat)
head(result[, grep("score|barrier", names(result), ignore.case = TRUE)])

}
\seealso{
Other xfm paw: 
\code{\link{fpr_xfm_paw_barrier_result}()},
\code{\link{fpr_xfm_paw_score_cv_len}()},
\code{\link{fpr_xfm_paw_score_cv_slope}()},
\code{\link{fpr_xfm_paw_score_embed}()},
\code{\link{fpr_xfm_paw_score_final}()},
\code{\link{fpr_xfm_paw_score_outlet_drop}()},
\code{\link{fpr_xfm_paw_score_swr}()},
\code{\link{fpr_xfm_paw_swr}()},
\code{\link{fpr_xfm_paw_xing_fix_size}()}
}
\concept{xfm paw}
