% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_xfm_paw_barrier_result.R
\name{fpr_xfm_paw_barrier_result}
\alias{fpr_xfm_paw_barrier_result}
\title{Transform PSCIS Assessment Worksheet Data by Calculating Barrier Result}
\usage{
fpr_xfm_paw_barrier_result(
  dat,
  col_crossing_type = crossing_type,
  col_final_score = final_score,
  col_barrier_result = barrier_result,
  risk_low_value = 15,
  risk_high_value = 19
)
}
\arguments{
\item{dat}{[dataframe] A dataframe containing the PSCIS data.}

\item{col_crossing_type}{[character] A column name specifying the crossing type. Default is `crossing_type`.}

\item{col_final_score}{[character] A column name specifying the final score. Default is `final_score`.}

\item{col_barrier_result}{[character] A column name for the output barrier result. Default is `barrier_result`.}

\item{risk_low_value}{[numeric] A numeric value representing the lower threshold for low-risk final scores. Default is `15`.}

\item{risk_high_value}{[numeric] A numeric value representing the upper threshold for moderate-risk final scores. Default is `19`.}
}
\value{
[dataframe] A dataframe with the specified column for barrier result added or updated.
}
\description{
This function calculates the barrier result for PSCIS data based on the crossing type and final score.
}
\examples{
dat <- data.frame(
  crossing_type = c("Open Bottom Structure", "Other", NA, "", "Other"),
  final_score = c(10, 16, 25, NA, 18)
)
fpr_xfm_paw_barrier_result(dat)

}
\seealso{
Other xfm paw: 
\code{\link{fpr_xfm_paw_score_cv_len}()},
\code{\link{fpr_xfm_paw_score_cv_slope}()},
\code{\link{fpr_xfm_paw_score_embed}()},
\code{\link{fpr_xfm_paw_score_final}()},
\code{\link{fpr_xfm_paw_score_outlet_drop}()},
\code{\link{fpr_xfm_paw_score_swr}()},
\code{\link{fpr_xfm_paw_swr}()}
}
\concept{xfm paw}
