% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_xfm_paw_score_cv_slope.R
\name{fpr_xfm_paw_score_cv_slope}
\alias{fpr_xfm_paw_score_cv_slope}
\title{Transform PSCIS Assessment Worksheet Data by Calculating Culvert Slope Score}
\usage{
fpr_xfm_paw_score_cv_slope(
  dat,
  col_culvert_slope_percent = culvert_slope_percent,
  col_culvert_slope_score = culvert_slope_score,
  slope_risk_low = 1,
  slope_risk_high = 3,
  risk_mod = 5,
  risk_high = 10,
  risk_low = 0
)
}
\arguments{
\item{dat}{[dataframe] A dataframe containing the PSCIS data.}

\item{col_culvert_slope_percent}{[character] A column name specifying the culvert slope percent, as a string or tidy-select syntax. Default is `culvert_slope_percent`.}

\item{col_culvert_slope_score}{[character] A column name for the output culvert slope score, as a string or tidy-select syntax. Default is `culvert_slope_score`.}

\item{slope_risk_low}{[numeric] A numeric value representing the lower threshold for low culvert slope risk (in percent). Default is `1`.}

\item{slope_risk_high}{[numeric] A numeric value representing the threshold for high culvert slope risk (in percent). Default is `3`.}

\item{risk_mod}{[numeric] A numeric value representing the risk score for moderate culvert slope conditions. Default is `5`.}

\item{risk_high}{[numeric] A numeric value representing the risk score for high culvert slope conditions. Default is `10`.}

\item{risk_low}{[numeric] A numeric value representing the risk score for low culvert slope conditions. Default is `0`.}
}
\value{
[dataframe] A dataframe with the specified column for culvert slope score added or updated.
}
\description{
This function calculates the culvert slope score for BC Provincial Stream Crossing Inventory System (PSCIS) data based on conditions involving culvert slope percent.
}
\examples{
dat <- data.frame(
  culvert_slope_percent = c(0.5, 1.5, 3.5, NA)
)
fpr_xfm_paw_score_cv_slope(
  dat
)

}
\seealso{
Other xfm paw: 
\code{\link{fpr_xfm_paw_all_scores_result}()},
\code{\link{fpr_xfm_paw_barrier_result}()},
\code{\link{fpr_xfm_paw_score_cv_len}()},
\code{\link{fpr_xfm_paw_score_embed}()},
\code{\link{fpr_xfm_paw_score_final}()},
\code{\link{fpr_xfm_paw_score_outlet_drop}()},
\code{\link{fpr_xfm_paw_score_swr}()},
\code{\link{fpr_xfm_paw_swr}()}
}
\concept{xfm paw}
