% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_xfm_paw_score_final.R
\name{fpr_xfm_paw_score_final}
\alias{fpr_xfm_paw_score_final}
\title{Transform PSCIS Assessment Worksheet Data by Calculating Final Score}
\usage{
fpr_xfm_paw_score_final(
  dat,
  col_final_score = final_score,
  col_culvert_length_score = culvert_length_score,
  col_embed_score = embed_score,
  col_outlet_drop_score = outlet_drop_score,
  col_culvert_slope_score = culvert_slope_score,
  col_stream_width_ratio_score = stream_width_ratio_score
)
}
\arguments{
\item{dat}{[dataframe] A dataframe containing the score data.}

\item{col_final_score}{[character] A column name for the final score output, as a string or tidy-select syntax. Default is `final_score`.}

\item{col_culvert_length_score}{[character] A column name for the culvert length score, as a string or tidy-select syntax. Default is `culvert_length_score`.}

\item{col_embed_score}{[character] A column name for the embed score, as a string or tidy-select syntax. Default is `embed_score`.}

\item{col_outlet_drop_score}{[character] A column name for the outlet drop score, as a string or tidy-select syntax. Default is `outlet_drop_score`.}

\item{col_culvert_slope_score}{[character] A column name for the culvert slope score, as a string or tidy-select syntax. Default is `culvert_slope_score`.}

\item{col_stream_width_ratio_score}{[character] A column name for the stream width ratio score, as a string or tidy-select syntax. Default is `stream_width_ratio_score`.}
}
\value{
[dataframe] A dataframe with the specified column for the final score added or updated.
}
\description{
This function calculates the final score for BC Provincial Stream Crossing Inventory System (PSCIS) data by summing specified score columns.
}
\examples{
dat <- data.frame(
  culvert_length_score = c(1, 2, 3, NA),
  embed_score = c(2, 3, NA, 1),
  outlet_drop_score = c(3, NA, 2, 1),
  culvert_slope_score = c(4, 3, 2, NA),
  stream_width_ratio_score = c(5, NA, 4, 3)
)
fpr_xfm_paw_score_final(
  dat
)

}
\seealso{
Other xfm paw: 
\code{\link{fpr_xfm_paw_all_scores_result}()},
\code{\link{fpr_xfm_paw_barrier_result}()},
\code{\link{fpr_xfm_paw_score_cv_len}()},
\code{\link{fpr_xfm_paw_score_cv_slope}()},
\code{\link{fpr_xfm_paw_score_embed}()},
\code{\link{fpr_xfm_paw_score_outlet_drop}()},
\code{\link{fpr_xfm_paw_score_swr}()},
\code{\link{fpr_xfm_paw_swr}()}
}
\concept{xfm paw}
