% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_xfm_paw_score_outlet_drop.R
\name{fpr_xfm_paw_score_outlet_drop}
\alias{fpr_xfm_paw_score_outlet_drop}
\title{Transform PSCIS Assessment Worksheet Data by Calculating Outlet Drop Score}
\usage{
fpr_xfm_paw_score_outlet_drop(
  dat,
  col_outlet_drop_meters = outlet_drop_meters,
  col_outlet_drop_score = outlet_drop_score,
  drop_risk_low = 0.15,
  drop_risk_high = 0.3,
  risk_mod = 5,
  risk_high = 10,
  risk_low = 0
)
}
\arguments{
\item{dat}{[dataframe] A dataframe containing the PSCIS data.}

\item{col_outlet_drop_meters}{[character] A column name specifying the outlet drop meters, as a string or tidy-select syntax. Default is `outlet_drop_meters`.}

\item{col_outlet_drop_score}{[character] A column name for the output outlet drop score, as a string or tidy-select syntax. Default is `outlet_drop_score`.}

\item{drop_risk_low}{[numeric] A numeric value representing the lower threshold for low outlet drop risk. Default is `0.15`.}

\item{drop_risk_high}{[numeric] A numeric value representing the threshold for high outlet drop risk. Default is `0.30`.}

\item{risk_mod}{[numeric] A numeric value representing the risk score for moderate outlet drop conditions. Default is `5`.}

\item{risk_high}{[numeric] A numeric value representing the risk score for high outlet drop conditions. Default is `10`.}

\item{risk_low}{[numeric] A numeric value representing the risk score for low outlet drop conditions. Default is `0`.}
}
\value{
[dataframe] A dataframe with the specified column for outlet drop score added or updated.
}
\description{
This function calculates the outlet drop score for PSCIS data based on conditions involving outlet drop meters.
}
\examples{
dat <- data.frame(
  outlet_drop_meters = c(0.1, 0.2, 0.35, NA)
)
fpr_xfm_paw_score_outlet_drop(
  dat
)

}
\seealso{
Other xfm paw: 
\code{\link{fpr_xfm_paw_all_scores_result}()},
\code{\link{fpr_xfm_paw_barrier_result}()},
\code{\link{fpr_xfm_paw_score_cv_len}()},
\code{\link{fpr_xfm_paw_score_cv_slope}()},
\code{\link{fpr_xfm_paw_score_embed}()},
\code{\link{fpr_xfm_paw_score_final}()},
\code{\link{fpr_xfm_paw_score_swr}()},
\code{\link{fpr_xfm_paw_swr}()},
\code{\link{fpr_xfm_paw_xing_fix_size}()}
}
\concept{xfm paw}
