% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_xfm_paw_swr.R
\name{fpr_xfm_paw_swr}
\alias{fpr_xfm_paw_swr}
\title{Transform PSCIS Assessment Worksheet Data by Calculating Stream Width Ratio}
\usage{
fpr_xfm_paw_swr(
  dat,
  col_downstream_channel_width_meters = downstream_channel_width_meters,
  col_diameter_or_span_meters = diameter_or_span_meters,
  digits = 1
)
}
\arguments{
\item{dat}{A dataframe containing the PSCIS data.}

\item{col_downstream_channel_width_meters}{Column name for downstream channel width, as a string or tidy-select syntax. Default is `downstream_channel_width_meters`.}

\item{col_diameter_or_span_meters}{Column name for diameter or span, as a string or tidy-select syntax. Default is `diameter_or_span_meters`.}

\item{digits}{Passed to `[round()]` function to specify the number of decimal places in the output. The default value is `1`.}
}
\value{
A dataframe with `stream_width_ratio` column added or updated.
}
\description{
This function calculates the stream width ratio (SWR) for PSCIS data based on the ratio of two specified columns: one for downstream channel width and one for diameter or span.
}
\examples{
dat <- data.frame(
  downstream_channel_width_meters = c(5, NA, 10),
  diameter_or_span_meters = c(2, 3, NA)
)
fpr_xfm_paw_score_swr(
  dat,
  col_downstream_channel_width_meters = downstream_channel_width_meters,
  col_diameter_or_span_meters = diameter_or_span_meters
)

}
\seealso{
Other xfm paw: 
\code{\link{fpr_xfm_paw_barrier_result}()},
\code{\link{fpr_xfm_paw_score_cv_len}()},
\code{\link{fpr_xfm_paw_score_cv_slope}()},
\code{\link{fpr_xfm_paw_score_embed}()},
\code{\link{fpr_xfm_paw_score_final}()},
\code{\link{fpr_xfm_paw_score_outlet_drop}()},
\code{\link{fpr_xfm_paw_score_swr}()}
}
\concept{xfm paw}
