% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_xref_template_pscis.R
\name{fpr_xref_template_pscis}
\alias{fpr_xref_template_pscis}
\title{Create blank PSCIS submission template}
\usage{
fpr_xref_template_pscis()
}
\description{
For use in prepping digital field data for submission.
}
\examples{
example <- fpr_xref_template_pscis()
}
