% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfpr_create_hydrograph.R
\name{fpr_create_hydrograph}
\alias{fpr_create_hydrograph}
\title{Creates hydrographs}
\usage{
fpr_create_hydrograph(station, pane_hydat = TRUE, single_hydat = TRUE)
}
\arguments{
\item{station}{String (quoted) number of station}

\item{pane_hydat}{Boolean TRUE if you want a pane layout of all hydrographs}

\item{single_hydat}{Boolean TRUE if you want a single hydrograph with mean flows}
}
\value{
Hydrographs of station
}
\description{
Creates hydrographs
}
\examples{
\dontrun{fpr_create_hydrograph('08EE004', pane_hydat = FALSE))}
}
