% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_db_conn.R
\name{fpr_db_conn}
\alias{fpr_db_conn}
\title{Connect to postgres database}
\usage{
fpr_db_conn(
  db_var = "PG_DB_SHARE",
  host_var = "PG_HOST_SHARE",
  port_var = "PG_PORT_SHARE",
  user_var = "PG_USER_SHARE",
  password_var = "PG_PASS_SHARE"
)
}
\arguments{
\item{db_var}{Quoted string value of environmental variable with database name}

\item{host_var}{Quoted string value of environmental variable with host name}

\item{port_var}{Quoted string value of environmental variable with port name}

\item{user_var}{Quoted string value of environmental variable with user name}

\item{password_var}{Quoted string value of environmental variable with password}
}
\description{
Connect to pg database using environmental variables
}
\examples{
\dontrun{fpr_db_conn()}
}
