% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_db_conn.R
\name{fpr_db_conn}
\alias{fpr_db_conn}
\title{Connect to postgres database}
\usage{
fpr_db_conn(
  db_var = Sys.getenv("PG_DB_SHARE"),
  host_var = Sys.getenv("PG_HOST_SHARE"),
  port_var = Sys.getenv("PG_PORT_SHARE"),
  user_var = Sys.getenv("PG_USER_SHARE"),
  password_var = Sys.getenv("PG_PASS_SHARE")
)
}
\arguments{
\item{db_var}{Quoted character string value of database name.}

\item{host_var}{Quoted character string value of host name.}

\item{port_var}{Quoted character string value of port name.}

\item{user_var}{Quoted character string value of user name.}

\item{password_var}{Quoted character string value of password.}
}
\description{
Connect to postgres database
}
\examples{
\dontrun{conn <- fpr_db_conn()}
}
\seealso{
Other database postgres: 
\code{\link{fpr_db_query}()},
\code{\link{fpr_dbq_lscols}()},
\code{\link{fpr_dbq_lstables}()},
\code{\link{fpr_dbq_subset}()}
}
\concept{database postgres}
