% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_db_query.R
\name{fpr_db_query}
\alias{fpr_db_query}
\title{Query a postgres database}
\usage{
fpr_db_query(query = "SELECT * FROM bcfishpass.crossings limit 100000;")
}
\arguments{
\item{query}{Quoted string query written in `sql`.  Defaults to  "SELECT * FROM bcfishpass.crossings limit 100000;"}
}
\value{
Object pulled from database.
}
\description{
Provides connection from \link{tfpr_db_conn} as well as a subsequent disconnection
}
\examples{
\dontrun{fpr_db_query(query = fpr_db_q_crossings_utm())}
}
