% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fpr_import_pscis}
\alias{fpr_import_pscis}
\title{Import the pscis template}
\usage{
fpr_import_pscis(workbook_name = "pscis_phase1.xlsm")
}
\arguments{
\item{workbook_name}{string value for name of worksheet}
}
\value{
A dataframe
}
\description{
Import the pscis template.  This is a helper for fpr_import_all
}
\examples{
fpr_import_pscis()
}
