% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fpr_kable}
\alias{fpr_kable}
\title{Custom kable settings}
\usage{
fpr_kable(
  dat,
  caption_text = "",
  font = font_set,
  footnote_text = NULL,
  scroll = TRUE,
  scroll_box_height = "500px",
  col_width_min = NULL,
  width_min = "1.0in"
)
}
\arguments{
\item{dat}{dataframe to make a table with}

\item{caption_text}{string to insert as caption}

\item{font}{size of font.  Defaults to font_set usually set in setup chunks of index.Rmd based on output type ie. gitbook vs pagedown}

\item{footnote_text}{string to insert as footnote}

\item{scroll}{TRUE or FALSE about whether to have scroll}

\item{scroll_box_height}{string. pixel height of scroll box.  Defaults to "500px"}

\item{col_width_min}{Number or vector of numbers. Defaults to NULL}

\item{width_min}{string with units of inches (in). Defaults to 1.0in}
}
\description{
Custom kable settings
}
\examples{
fpr_kable(mtcars)
}
