% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{fpr_make_geopackage}
\alias{fpr_make_geopackage}
\title{Make geopackage in directory called}
\usage{
fpr_make_geopackage(
  dat,
  dir = "data/fishpass_mapping/",
  gpkg_name = "fishpass_mapping",
  utm_zone = 9,
  x = "utm_easting",
  y = "utm_northing",
  crs = 4326
)
}
\arguments{
\item{dat}{Dataframe with coordinates to burn to geopackage}

\item{dir}{String (quoted) representing relative path to directory where geopackage will be burned}

\item{gpkg_name}{String Name of the geopackage}

\item{utm_zone}{Integer UTM zone}

\item{x}{String (quoted) representing column name where x coordinates are stored}

\item{y}{String (quoted) representing column name where y coordinates are stored}

\item{crs}{Integer destination CRS. Defaults to 4326}
}
\value{
geopackage layer
}
\description{
Names the layer the same as the name of the dataframe that is burned
}
