% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_bcfishpass}
\alias{fpr_my_bcfishpass}
\title{Filter the bcfishpass.crossings table export by PSCIS stream_crossing_id}
\usage{
fpr_my_bcfishpass(
  dat = bcfishpass,
  site = my_site,
  col_filter = stream_crossing_id,
  col_pull = stream_order,
  round_dig = 0
)
}
\arguments{
\item{dat}{Dataframe. Defaults to bcfishpass filtered by Phase 2 sites only}

\item{site}{Numeric PSCIS stream_crossing_id. Defaults to value defined by my_site}

\item{round_dig}{Integer for how many numbers to round to. Defaults to 0.}

\item{col_to_filter}{String value of column to filter}

\item{col_to_pull}{String vlaue of column to pull}
}
\value{
Vector
}
\description{
Filter the bcfishpass.crossings table export by PSCIS stream_crossing_id
}
