% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fpr_my_fish_sp}
\alias{fpr_my_fish_sp}
\title{#' Pull out fish species long form names from our fish species codes in bcfishpass export
#'
#' @param ... Parameters to pass to fpr_my_bcfishpass
#'
#' @return Vector
#' @export
#'
#' @examples
fpr_my_fish_sp <- function(...){
  # col_to_pull <-  sym(col_to_pull)
  str_to_pull <- stringr::str_replace_all(
    (fpr_my_bcfishpass(..., col_to_pull = observedspp_upstr)),
     c("\\{" = "","\\}" = "")) %>%
    strsplit(., ",") %>%
    unlist()
  fishbc::freshwaterfish %>%
    dplyr::filter(Code %in% str_to_pull &
             !Code %in% c('SST','TR')) %>%
    dplyr::pull(CommonName) %>%
    stringr::str_to_lower() %>%
    knitr::combine_words()
}
Pull out fish species long form names from our fish species codes in bcfishpass export}
\usage{
fpr_my_fish_sp(
  ...,
  col_pull = observedspp_upstr,
  sp_to_ignore = c("SST", "TR")
)
}
\arguments{
\item{...}{Parameters to pass to fpr_my_bcfishpass}

\item{col_pull}{Name of column to pull from bcfishpass. Defaults to observedspp_upstr. observedspp_dnstr is the other option}

\item{sp_to_ignore}{Vector of strings representing species codes to ignore in output. Defaults to SST and TR as these seem a bit ambigous.}
}
\value{
Vector
}
\description{
#' Pull out fish species long form names from our fish species codes in bcfishpass export
#'
#' @param ... Parameters to pass to fpr_my_bcfishpass
#'
#' @return Vector
#' @export
#'
#' @examples
fpr_my_fish_sp <- function(...){
  # col_to_pull <-  sym(col_to_pull)
  str_to_pull <- stringr::str_replace_all(
    (fpr_my_bcfishpass(..., col_to_pull = observedspp_upstr)),
     c("\\{" = "","\\}" = "")) %>%
    strsplit(., ",") %>%
    unlist()
  fishbc::freshwaterfish %>%
    dplyr::filter(Code %in% str_to_pull &
             !Code %in% c('SST','TR')) %>%
    dplyr::pull(CommonName) %>%
    stringr::str_to_lower() %>%
    knitr::combine_words()
}
Pull out fish species long form names from our fish species codes in bcfishpass export
}
