% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_habitat_paragraph}
\alias{fpr_my_habitat_paragraph}
\title{Paragraph to describe the cover, widths, gradient and substrates.}
\usage{
fpr_my_habitat_paragraph(loc = "us", sit = my_site)
}
\arguments{
\item{loc}{String in quotes which is either 'us' upstream or 'ds' downstream.  Defaults to 'us'}

\item{sit}{Integer. Corresponds to the PSCIS site id. Defaults to my_site defined in environment}
}
\value{
String sentence with the cover types described
}
\description{
Randomly generated order from the inputs of cover, channel and substrate sentences.
}
\examples{
fpr_my_cover_sentence()
}
