% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{fpr_my_wshd}
\alias{fpr_my_wshd}
\title{PUll out watershed characteristics}
\usage{
fpr_my_wshd(site = my_site, col = area_km)
}
\arguments{
\item{site}{Integer - defaults to my_site}

\item{col}{string unquoted representing column to pull from wshds object}
}
\description{
PUll out watershed characteristics
}
