% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_document_all}
\alias{fpr_photo_document_all}
\title{Title}
\usage{
fpr_photo_document_all(path_to_photo_dir, full_names = T)
}
\arguments{
\item{path_to_photo_dir}{Full path to photo directories to be}

\item{full_names}{}
}
\value{

}
\description{
Title
}
