% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_paths_to_copy}
\alias{fpr_photo_paths_to_copy}
\title{Identify photos that should be copied over into file to upload to PSCIS}
\usage{
fpr_photo_paths_to_copy(path_to_photo_dir)
}
\arguments{
\item{path_to_photo_dir}{Full path to photo directories that will be copied over}
}
\value{

}
\description{
Identify photos that should be copied over into file to upload to PSCIS
}
