% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_qa}
\alias{fpr_photo_qa}
\title{Helper function to QA Photos}
\usage{
fpr_photo_qa(dat = pscis_all, col = site_id, dir_photos = "data/photos/")
}
\arguments{
\item{dat}{Dataframe that contains a column with the numeric site ids. Defaults to `pscis_all`}

\item{col}{Column to pull to get site IDs. Defaults to site_id}

\item{dir_photos}{Directory that contains the photos. Defaults to data/photos/}
}
\value{
List of dataframes with photo names for 6 PSCIS photos. Will return NAs when photo not there and list of
names when there is more than one that matches the pattern.
}
\description{
Ensure that there are 6 named photos and no duplicates named.
Used inside of \link{fpr_photo_qa_df}.
}
