% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_qa}
\alias{fpr_photo_qa}
\title{Helper function to QA Photos}
\usage{
fpr_photo_qa(dat = pscis_all, col = site_id)
}
\arguments{
\item{dat}{Dataframe to pull site IDs from that coincide with photo folders. Defaults to pscis_phase1}

\item{col}{Column to pull to get site IDs. Defaults to my_crossing_reference}
}
\value{
List of tibbles for each sites.
}
\description{
Ensure that there are 6 named photos and no duplicates named.  Uses \link{fpr_photo_qa_prep}.
Used inside of \link{fpr_photo_qa_df}.
}
