% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_qa_df}
\alias{fpr_photo_qa_df}
\title{QA photos to see if all required PSCIS photos are named and look for duplicates of those names.}
\usage{
fpr_photo_qa_df(...)
}
\arguments{
\item{...}{Not used. Use to pass string path defined as `dir_photos = 'quoted_filepath'` to run
QA outside of defualt filepath of root of helper function  \link{fpr_photo_qa}.}
}
\description{
QA photos to see if all required PSCIS photos are named and look for duplicates of those names.
}
