% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_qa_missing_all}
\alias{fpr_photo_qa_missing_all}
\title{Find directories that exist but have none of the required 6 PSICS photos tagged within their name.
Fed into \link{fpr_photo_qa_df} and relies on \link{fpr_photo_qa}.  Site directories must be named as numbers.}
\usage{
fpr_photo_qa_missing_all(...)
}
\arguments{
\item{...}{Not used. Used to pass `dir_photos = 'filepath'` to \link{fpr_photo_qa}}
}
\value{
Tibble with 7 columns.
}
\description{
Find directories that exist but have none of the required 6 PSICS photos tagged within their name.
Fed into \link{fpr_photo_qa_df} and relies on \link{fpr_photo_qa}.  Site directories must be named as numbers.
}
