% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_qa_prep}
\alias{fpr_photo_qa_prep}
\title{Prep for fpr_photo_qa which will QA photos}
\usage{
fpr_photo_qa_prep(site_id)
}
\arguments{
\item{site_id}{Numeric value of site corresponding to folder name}
}
\value{

}
\description{
Ensure that there are 6 named photos and no duplicates named
}
