% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{fpr_photo_rotate}
\alias{fpr_photo_rotate}
\title{Rotate an image}
\usage{
fpr_photo_rotate(
  site_id = my_site,
  dir_photos = "data/photos/",
  rotate = 180,
  str_to_pull = "barrel",
  str_prepend = "rotated_",
  ...
)
}
\arguments{
\item{site_id}{Numeric indicating site id cooresponding to name of directory containing the images}

\item{dir_photos}{String quoted for directory where photo directories are kept.  Trailing forward slash required.
Defaults to 'data/photos/'.}

\item{rotate}{Numeric value of the amount of rotation in degrees. Defaults to 180}

\item{str_to_pull}{Sting value in photo name.  Must be unique in file}

\item{...}{Not used.  Used to pass `dir_photos` to \link{fpr_photo_pull_by_str}}
}
\value{
Rotated photo burned into directory of targeted original.
}
\description{
Rotate an image
}
