% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fpr_pscis_wkb_paths}
\alias{fpr_pscis_wkb_paths}
\title{Get Names of PSCIS Files}
\usage{
fpr_pscis_wkb_paths()
}
\value{
String of pull path name of file by that pattern
}
\description{
Reads filenames in your data file for files with the 'pscis' string in their name. Ignores temp files that are there when
}
