% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{fpr_sheet_trim}
\alias{fpr_sheet_trim}
\title{Trim up an excel worksheet - usually during import}
\usage{
fpr_sheet_trim(dat)
}
\arguments{
\item{dat}{Dataframe}
}
\value{
A dataframe
}
\description{
Trim up a dataframe based on complete rows. Tidy header names
}
