% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fpr_sheet_trim}
\alias{fpr_sheet_trim}
\title{Trim up an excel worksheet - usually during import}
\usage{
fpr_sheet_trim(dat)
}
\arguments{
\item{dat}{string value for name of worksheet}
}
\value{
A dataframe
}
\description{
Trim up a dataframe based on complete rows and col. Tidy header names
}
\examples{
fpr_import_pscis()
}
