% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{fpr_sp_wshd_stats}
\alias{fpr_sp_wshd_stats}
\title{Get elevation statistics}
\usage{
fpr_sp_wshd_stats(dat = wshds)
}
\arguments{
\item{dat}{sf object of polygons to gather elevation statistics for}
}
\value{
sf object of polygons with elevation information added
}
\description{
Get elevation statistics
}
