% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_xfm_paw_score_cv_len.R
\name{fpr_xfm_paw_score_cv_len}
\alias{fpr_xfm_paw_score_cv_len}
\title{Transform PSCIS Assessment Worksheet Data by Calculating Culvert Length Score}
\usage{
fpr_xfm_paw_score_cv_len(
  dat,
  col_length_or_width_meters = length_or_width_meters
)
}
\arguments{
\item{dat}{A dataframe containing the PSCIS data.}

\item{col_length_or_width_meters}{Column name for length or width, as a string or tidy-select syntax. Default is `length_or_width_meters`.}
}
\value{
A dataframe with `culvert_length_score` column added or updated.
}
\description{
This function calculates the culvert length score for PSCIS data based on the given length or width column.
}
\examples{
dat <- data.frame(
  length_or_width_meters = c(10, 20, 35, NA)
)
fpr_xfm_culvert_length_score(
  dat,
  col_length_or_width_meters = length_or_width_meters
)

}
\seealso{
Other xfm paw: 
\code{\link{fpr_xfm_paw_score_embed}()},
\code{\link{fpr_xfm_paw_swr}()}
}
\concept{xfm paw}
