% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_xfm_paw_score_embed.R
\name{fpr_xfm_paw_score_embed}
\alias{fpr_xfm_paw_score_embed}
\title{Transform PSCIS Assessment Worksheet Data by Calculating Embedment Score}
\usage{
fpr_xfm_paw_score_embed(
  dat,
  col_continuous_embeddedment_yes_no = continuous_embeddedment_yes_no,
  col_diameter_or_span_meters = diameter_or_span_meters,
  col_average_depth_embededdment_meters = average_depth_embededdment_meters
)
}
\arguments{
\item{dat}{A dataframe containing the PSCIS data.}

\item{col_continuous_embeddedment_yes_no}{Column name for continuous embedment ("Yes" or "No"), as a string or tidy-select syntax. Default is `continuous_embeddedment_yes_no`.}

\item{col_diameter_or_span_meters}{Column name for diameter or span, as a string or tidy-select syntax. Default is `diameter_or_span_meters`.}

\item{col_average_depth_embededdment_meters}{Column name for average depth of embedment, as a string or tidy-select syntax. Default is `average_depth_embededdment_meters`.}
}
\value{
A dataframe with `embed_score` column added or updated.
}
\description{
This function calculates the embedment score for PSCIS data based on conditions involving continuous embedment, pipe diameter, and average depth of embedment.
}
\examples{
dat <- data.frame(
  continuous_embeddedment_yes_no = c("No", "Yes", "Yes"),
  diameter_or_span_meters = c(2, NA, 3),
  average_depth_embededdment_meters = c(0.4, 0.1, NA)
)
fpr_xfm_paw_score_embed(
  dat
)

}
\seealso{
Other xfm paw: 
\code{\link{fpr_xfm_paw_score_cv_len}()},
\code{\link{fpr_xfm_paw_swr}()}
}
\concept{xfm paw}
